Added upon the new multiple categories
Entries have been added not in response to seeing a related model, but rather as preemptive measures to prevent needing to block models in the future
Added upon entries in the furry and anime categories. 
Created Pokemon Category
Note: some pokemon show up in the furry lists. Namely, anthro pokemon.